Move multimodal HTML viewer code to a separate html_viewer module
